Renames useFormRef
