+++ src/useFormRef/index.ts
@@ -1,1 +1,1 @@
-export { useAnchor } from './useAnchor';
+export { useAnchor } from './useFormRef';
